+++ notifications/CHANGELOG.md
@@ -4,10 +4,11 @@ This file documents changes to the `workbench-notifications` library, including 
 
 ## 0.3
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.3-d74ff96"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.3-89d0d9e"`
 
 ### Changed
 - Moved `SamModel`
+- Target java 11
 
 ### Added
 - Support scala 2.13